+++ index.native.js
@@ -1,4 +1,5 @@
 import { View } from 'react-native';
+import PropTypes from 'prop-types';
 import React from 'react';
 import RNSvg, {
   Circle,
@@ -9,31 +10,44 @@ import RNSvg, {
   Image,
   Line,
   LinearGradient,
+  Mask,
   Path,
+  Pattern,
   Polygon,
   Polyline,
   RadialGradient,
   Rect,
   Stop,
   Symbol,
-  Text,
   TSpan,
+  Text,
   TextPath,
   Use
 } from 'react-native-svg';
 
+/**
+ * Return a SVG element.
+ *
+ * @param {Object} props The properties that are spread on the SVG element.
+ * @returns {React.Component} SVG.
+ * @public
+ */
 const Svg = (props) => {
   const { title, ...rest } = props;
 
   if (title) {
     return (
       <View accessible={ true } accessibilityLabel={ title }>
-        <RNSvg { ...rest} />
+        <RNSvg { ...rest } />
       </View>
     );
   }
 
-  return <RNSvg { ...rest} />
+  return <RNSvg { ...rest } />;
+};
+
+Svg.propTypes = {
+  title: PropTypes.string
 };
 
 export {
@@ -45,7 +59,9 @@ export {
   Image,
   Line,
   LinearGradient,
+  Mask,
   Path,
+  Pattern,
   Polygon,
   Polyline,
   RadialGradient,
@@ -53,8 +69,8 @@ export {
   Stop,
   Svg,
   Symbol,
-  Text,
   TSpan,
+  Text,
   TextPath,
   Use
 };